// webpack---_wix_analytics_ng_bm/constants/enums.ts
export enum EnterReportOriginBI {
  changeView = 'change view',
  drillDown = 'drill down',
  search = 'search',
  myReports = 'my reports',
  dashboardWidget = 'dashboard widget',
  insightsWidget = 'insights widget',
  subscribeEmail = 'subscribe email',
  alertEmail = 'alert email',
  sidebar = 'sidebar',
  analyticsMarketingPage = 'analytics marketing page',
  newReleases = 'new releases',
  insights = 'insights',
  suggestionsAnalyticsDigestEmail = 'suggestions analytics digest email',
  digestMailHowPeopleFindYou = 'digest mail how people find you',
  digestMailTrafficInsights = 'digest mail traffic insights',
  digestMailBookingsStats = 'digest mail bookings stats',
  digestMailStoreStats = 'digest mail store stats',
  digestMailBlogStats = 'digest mail blog stats',
  digestMailMainCTATop = 'digest mail main CTA top',
  digestMailVisitorsAndReferralTraffic = 'digest mail visitors and referral traffic',
  digestMailMainCTABottom = 'digest mail main CTA bottom',
  emailUpdates = 'email_updates',
  trafficOverviewPage = 'traffic overview page',
  salesOverviewPage = 'sales overview page',
  forecastPage = 'forecast page',
}
// ...
export enum AlertColorScheme {
  sales = 'sales',
  traffic = 'traffic',
  blog = 'blog',
  bookings = 'bookings',
  people = 'people',
}

export enum ReportCategories {
  accounting = 'category.title.accounting',
  blog = 'category.title.blog',
  booking = 'category.title.booking',
  customersAndUsers = 'category.title.customers.and.users',
  marketing = 'category.title.marketing',
  revenuesOrSales = 'category.title.revenues.or.sales',
  store = 'category.title.store',
  subscriptions = 'category.title.subscriptions',
  traffic = 'category.title.traffic',
  visitorsBehaviour = 'category.title.visitors.behaviour',
}